02-alphazero-chess/pyproject.toml
[project]
name = "alphazero-chess"
version = "0.1.0"
description = "AlphaZero on chess — small-scale, learning-first"
requires-python = ">=3.12,<3.13"
dependencies = [
    "torch>=2.5",
    "numpy>=1.26",
    "python-chess>=1.999",
    "tqdm>=4.66",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/alphazero"]
